+++ app/src/main/java/com/codesoom/assignment/AppHandler.java
@@ -10,7 +10,10 @@ public class AppHandler extends CommonHandler {
     @Override
     public void handle(HttpExchange exchange) throws IOException {
         String path = exchange.getRequestURI().getPath();
-        if(path.startsWith("/tasks")){
+        if(path.startsWith("/tasks/")){
+            TaskByIdHandler taskByIdHandler = new TaskByIdHandler();
+            taskByIdHandler.handle(exchange);
+        }else if(Objects.equals(path, "/tasks")){
             TaskHandler taskHandler = new TaskHandler();
             taskHandler.handle(exchange);
         }else if(Objects.equals(path, "/")){
