+++ app/src/main/java/com/codesoom/assignment/TaskHandler.java
@@ -11,32 +11,7 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 public class TaskHandler extends CommonHandler {
-    private ArrayList<Task> tasks = new ArrayList<Task>();
-    private ObjectMapper objectMapper = new ObjectMapper();
-
-    private String taskToJson(Task task) {
-        String jsonString = null;
-        try {
-            jsonString = objectMapper.writerWithDefaultPrettyPrinter().writeValueAsString(task);
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-            System.out.println("e: " + e);
-        }
-        System.out.println("JSON string: " + jsonString);
-        return jsonString;
-    }
-
-    private Task jsonToTask(String json) {
-        Task task = null;
-        try {
-            task = objectMapper.readValue(json, Task.class);
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-            System.out.println("e: " + e);
-        }
-        System.out.println("Task: " + task);
-        return task;
-    }
+    private static ArrayList<Task> tasks = new ArrayList<Task>();
 
     @Override
     public void handle(HttpExchange exchange) throws IOException {
@@ -46,15 +21,12 @@ public class TaskHandler extends CommonHandler {
             handleGetMethod(exchange);
         }else if(Objects.equals(method, "POST")){
             handlePostMethod(exchange);
-        }else if(Objects.equals(method, "PUT") || Objects.equals(method, "PATCH")){
-            handlePutOrPatchMethod(exchange);
-        }else if(Objects.equals(method, "DELETE")){
-            handleDeleteMethod(exchange);
         }
     }
 
     private void handleGetMethod(HttpExchange exchange) throws IOException {
-        String content = "Task: GET";
+        TaskSerializer taskSerializer = new TaskSerializer();
+        String content = taskSerializer.tasksToJson(tasks);
         exchange.sendResponseHeaders(200, content.getBytes().length);
         outputResponse(exchange, content);
     }
@@ -62,23 +34,12 @@ public class TaskHandler extends CommonHandler {
     private void handlePostMethod(HttpExchange exchange) throws IOException {
         InputStream inputStream = exchange.getRequestBody();
         String requestBody = new BufferedReader(new InputStreamReader(inputStream, "UTF-8")).lines().collect(Collectors.joining());
-        Task task = jsonToTask(requestBody);
+        TaskSerializer taskSerializer = new TaskSerializer();
+        Task task = taskSerializer.jsonToTask(requestBody);
         task.setId(tasks.size() + 1);
         tasks.add(task);
-        String content = taskToJson(task);
+        String content = taskSerializer.taskToJson(task);
         exchange.sendResponseHeaders(201, content.getBytes().length);
-        outputResponse(exchange, content);
-    }
-
-    private void handlePutOrPatchMethod(HttpExchange exchange) throws IOException {
-        String content = "Task: PUT or PATCH";
-        exchange.sendResponseHeaders(200, content.getBytes().length);
-        outputResponse(exchange, content);
-    }
-
-    private void handleDeleteMethod(HttpExchange exchange) throws IOException {
-        String content = "Task: DELETE";
-        exchange.sendResponseHeaders(200, content.getBytes().length);
         outputResponse(exchange, content);
     }
 }
